backend/requirements.txt
fastapi[all]==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
pydantic==1.10.9
openai==0.27.8
anthropic==0.3.6